[ POCRice - week2 ] - think I figured out the merge function
